+++ full-node/sov-ethereum/src/gas_price/cache.rs
@@ -4,16 +4,16 @@ use reth_primitives::B256;
 use reth_rpc_types::{Block, Rich};
 use schnellru::{ByLength, LruMap};
 use sov_evm::EthResult;
-use sov_modules_api::{DaSpec, WorkingSet};
+use sov_modules_api::WorkingSet;
 
 /// Block cache for gas oracle
-pub struct BlockCache<S: sov_modules_api::Spec, Da: DaSpec> {
+pub struct BlockCache<S: sov_modules_api::Spec> {
     cache: Mutex<LruMap<B256, Rich<Block>, ByLength>>,
-    provider: sov_evm::Evm<S, Da>,
+    provider: sov_evm::Evm<S>,
 }
 
-impl<S: sov_modules_api::Spec, Da: DaSpec> BlockCache<S, Da> {
-    pub fn new(max_size: u32, provider: sov_evm::Evm<S, Da>) -> Self {
+impl<S: sov_modules_api::Spec> BlockCache<S> {
+    pub fn new(max_size: u32, provider: sov_evm::Evm<S>) -> Self {
         Self {
             cache: Mutex::new(LruMap::new(ByLength::new(max_size))),
             provider,