+++ full-node/sov-ethereum/src/gas_price/gas_oracle.rs
@@ -8,7 +8,7 @@ use reth_primitives::{B256, U256, U64};
 use reth_rpc_types::BlockTransactions;
 use serde::{Deserialize, Serialize};
 use sov_evm::{EthApiError, EthResult, Evm, RpcInvalidTransactionError};
-use sov_modules_api::{DaSpec, WorkingSet};
+use sov_modules_api::WorkingSet;
 use tokio::sync::Mutex;
 use tracing::warn;
 
@@ -86,20 +86,20 @@ impl GasPriceOracleConfig {
 }
 
 /// Calculates a gas price depending on recent blocks.
-pub struct GasPriceOracle<S: sov_modules_api::Spec, Da: DaSpec> {
+pub struct GasPriceOracle<S: sov_modules_api::Spec> {
     /// The type used to get block and tx info
-    provider: Evm<S, Da>,
+    provider: Evm<S>,
     /// The config for the oracle
     oracle_config: GasPriceOracleConfig,
     /// The latest calculated price and its block hash
     last_price: Mutex<GasPriceOracleResult>,
     /// Cache
-    cache: BlockCache<S, Da>,
+    cache: BlockCache<S>,
 }
 
-impl<S: sov_modules_api::Spec, Da: DaSpec> GasPriceOracle<S, Da> {
+impl<S: sov_modules_api::Spec> GasPriceOracle<S> {
     /// Creates and returns the [GasPriceOracle].
-    pub fn new(provider: Evm<S, Da>, mut oracle_config: GasPriceOracleConfig) -> Self {
+    pub fn new(provider: Evm<S>, mut oracle_config: GasPriceOracleConfig) -> Self {
         // sanitize the percentile to be less than 100
         if oracle_config.percentile > 100 {
             warn!(prev_percentile = ?oracle_config.percentile, "Invalid configured gas price percentile, assuming 100");
@@ -112,7 +112,7 @@ impl<S: sov_modules_api::Spec, Da: DaSpec> GasPriceOracle<S, Da> {
             provider: provider.clone(),
             oracle_config,
             last_price: Default::default(),
-            cache: BlockCache::<S, Da>::new(max_header_history, provider),
+            cache: BlockCache::<S>::new(max_header_history, provider),
         }
     }